+++ src/main/java/com/vladbstrv/customer/CustomerRepository.java
@@ -3,9 +3,9 @@ package com.vladbstrv.customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository
-        extends JpaRepository<Customer, Integer> {
+        extends JpaRepository<Customer, Long> {
 
 
     boolean existsCustomerByEmail(String email);
-    boolean existsCustomerById(Integer id);
+    boolean existsCustomerById(Long id);
 }
